♻️ (dashboard): Move WRITE_SCOPES out of the GrantScopes component module
Exporting a constant alongside the component tripped
react-refresh/only-export-components. Keep GrantScopes.tsx component-only and
put the scope set in a plain module beside it.

refs AAASM-5180
