+++ dashboard/src/features/alerts/AlertRuleForm.test.tsx
@@ -6,7 +6,8 @@ import { AlertRuleForm } from './AlertRuleForm'
 import { ToastProvider } from '../../components/ToastProvider'
 import { AuthContext, type AuthContextValue, type Scope } from '../../auth/AuthContext'
 import type { AlertRule, Destination } from './types'
-import { GrantScopes, WRITE_SCOPES } from '../../auth/GrantScopes'
+import { GrantScopes } from '../../auth/GrantScopes'
+import { WRITE_SCOPES } from '../../auth/testScopes'
 
 // ── fetch stub mirroring the AAASM-1075 test setup ─────────────────────────
