+++ dashboard/src/features/alerts/AlertRulesTable.test.tsx
@@ -3,7 +3,8 @@ import userEvent from '@testing-library/user-event'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 import { AlertRulesTable } from './AlertRulesTable'
-import { GrantScopes, WRITE_SCOPES } from '../../auth/GrantScopes'
+import { GrantScopes } from '../../auth/GrantScopes'
+import { WRITE_SCOPES } from '../../auth/testScopes'
 import { ToastProvider } from '../../components/ToastProvider'
 import type { AlertRule } from './types'
