+++ dashboard/src/features/alerts/SilenceAction.test.tsx
@@ -5,7 +5,8 @@ import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 import { SilenceAction } from './SilenceAction'
 import { ToastProvider } from '../../components/ToastProvider'
 import { AuthContext, type AuthContextValue, type Scope } from '../../auth/AuthContext'
-import { GrantScopes, WRITE_SCOPES } from '../../auth/GrantScopes'
+import { GrantScopes } from '../../auth/GrantScopes'
+import { WRITE_SCOPES } from '../../auth/testScopes'
 
 interface Call {
   url: string
