+++ dashboard/src/features/liveOps/ApprovalPool.test.tsx
@@ -7,7 +7,8 @@ import { api } from '../../api/client'
 import { absent, known } from '../../lib/truthfulness'
 import { APPROVALS_QUERY_KEY, type Approval } from '../approvals/api'
 import { ApprovalPool } from './ApprovalPool'
-import { GrantScopes, WRITE_SCOPES } from '../../auth/GrantScopes'
+import { GrantScopes } from '../../auth/GrantScopes'
+import { WRITE_SCOPES } from '../../auth/testScopes'
 
 /** A real approval id: a UUID, which is what the decide endpoints parse. */
 const UUID_1 = '3f1c9a52-0c4e-4a1b-9f2d-6a7b8c9d0e1f'
